conformance/functionalityTesting/basics/createContext.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following conditions
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>createContext conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
try {
    debug("CreateContext functionality test case.");

    var wtu = WebCLTestUtils;
    var webCLPlatform = wtu.getPlatform();
    var webCLDevices = wtu.getDevices(webCLPlatform);
    var webCLContext, deviceType, contextDevices, contextProperty;

    var createContext = function(platform, devices, deviceType) {
        try {
            webCLContext = wtu.createContext(platform, devices, deviceType);

            contextDevices = webCLContext.getInfo(webcl.CONTEXT_DEVICES);
            contextProperty = webCLContext.getInfo(webcl.CONTEXT_PROPERTIES);

            if (typeof(platform) != 'undefined' && platform != null) {
                if (contextProperty.platform !== platform) {
                    console.log("Failed in validating Platform used to create WebCLContext. Expected : " + platform + ". Obtained : " + contextProperty.platform);
                    return null;
                }

                for (i = 0; i < contextDevices.length; i++)
                    if (contextDevices[i].getInfo(webcl.DEVICE_PLATFORM) !== platform) {
                        console.log("Failed in validating platform for devices[ " + i + "] used to create WebCLContext. Expected : " + platform + ". Obtained : " + contextDevices[i].getInfo(webcl.DEVICE_PLATFORM));
                        return null;
                    }
            }

            if (typeof(devices) != 'undefined' && devices != null) {
                for (i = 0; i < contextDevices.length; i++)
                    if (contextDevices[i] !== devices[i] ||
                            contextDevices[i].getInfo(webcl.DEVICE_PLATFORM) !== contextProperty.platform) {
                        console.log("Failed in validating devices[ " + i + " ] and its platform used to create WebCLContext. Expected : " + devices[i] + " && " + contextProperty.platform + ". Obtained : " + contextDevices[i] + " && " + contextDevices[i].getInfo(webcl.DEVICE_PLATFORM));
                        return null;
                    }
            }

            if (typeof(deviceType) != 'undefined' && deviceType != null) {
                // ALL and DEFAULT cannot be checked for. They are just alias to other types.
                if (deviceType == webcl.DEVICE_TYPE_ALL || deviceType == webcl.DEVICE_TYPE_DEFAULT)
                    return webCLContext;

                if (contextProperty.deviceType != deviceType) {
                    console.log("Failed in validating deviceType used to create WebCLContext. Expected : " + deviceType + ". Obtained : " + contextProperty.deviceType);
                    return null;
                }

                for (i = 0; i < contextDevices.length; i++)
                    if (contextDevices[i].getInfo(webcl.DEVICE_TYPE) != deviceType) {
                        console.log("Failed in validating deviceType for device[ " + i + " ] used to create WebCLContext. Expected : " + deviceType + ". Obtained : " + contextDevices[i].getInfo(webcl.DEVICE_TYPE));
                        return null;
                    }
            }
            return webCLContext;

        } catch(e) {
            e.description = "Exception thrown while creating and validating context : " + e.name;
            throw e;
        }
    }

    shouldBeTypeOrException("createContext()", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(webCLPlatform)", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(webCLPlatform, webCLDevices)", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(null, webCLDevices)", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(null, null, webcl.DEVICE_TYPE_CPU)", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(null, null, webcl.DEVICE_TYPE_GPU)", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(null, null, webcl.DEVICE_TYPE_ACCELERATOR)", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(null, null, webcl.DEVICE_TYPE_ALL)", "WebCLContext", "DEVICE_NOT_FOUND");
    shouldBeTypeOrException("createContext(null, null, webcl.DEVICE_TYPE_DEFAULT)", "WebCLContext", "DEVICE_NOT_FOUND");

} catch(e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
